feat(parser): reintroduce tags support
