--- examples/todotxt-to-json/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "todotxt-to-json"
-version = "0.1.0"
-authors = ["Zachary Golba <zachary.golba@postlight.com>"]
-edition = "2018"
-
-[dependencies]
-serde_json = "1.0"
-failure = "0.1"
-
-[dependencies.to2txt]
-features = ["serde"]
-path = "../../todotxt"
