--- examples/todotxt-to-json/README.md
@@ -1,30 +1,0 @@
-# todotxt-to-json
-
-## Installation
-
-```sh
-cargo install
-```
-
-## Usage
-
-```sh
-cat todo.txt | todotxt-to-json
-```
-
-## License
-
-Licensed under either of
-
-* Apache License, Version 2.0
-  ([LICENSE-APACHE](LICENSE-APACHE) or http://www.apache.org/licenses/LICENSE-2.0)
-* MIT license
-  ([LICENSE-MIT](LICENSE-MIT) or http://opensource.org/licenses/MIT)
-
-at your option.
-
-## Contribution
-
-Unless you explicitly state otherwise, any contribution intentionally submitted
-for inclusion in the work by you, as defined in the Apache-2.0 license, shall be
-dual licensed as above, without any additional terms or conditions.
